refactor: make nextauth options cleaner
