+++ src/modules/user/components/LoginForm/LoginForm.tsx
@@ -1,0 +1,2 @@
+const LoginForm = () => {};
+export default LoginForm;
