+++ src/utils/nextAuth/callbacks/session.ts
@@ -1,0 +1,11 @@
+import { CallbacksOptions } from 'next-auth';
+
+const session: CallbacksOptions['session'] = async ({
+    session,
+    user,
+    token,
+}) => {
+    session.user = token.user ?? session.user;
+    return session;
+};
+export default session;
