+++ src/utils/nextAuth/callbacks/signIn.ts
@@ -1,0 +1,17 @@
+import { CallbacksOptions } from 'next-auth';
+
+const signIn: CallbacksOptions['signIn'] = async ({
+    user,
+    account,
+    profile,
+    email,
+    credentials,
+}) => {
+    const isAllowedToSignIn = true;
+    if (isAllowedToSignIn) {
+        return true;
+    } else {
+        return false;
+    }
+};
+export default signIn;
